Fix XLSX generation
With help from Google Gemini
